Adds LCD support.
