--- main/buttons_modes_and_encoding_patterns_circuit.txt
@@ -1,1 +1,0 @@
-Add tinkercad circuit link here for the accompanying code.
