--- main/lcd.cpp
@@ -1,1 +1,0 @@
-// Some code here
